pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OBPAnalyser"
version = "0.0.1"
authors = [{ name = "Anton Wiberg", email = "wiberg.anton@gmail.com" }]
description = "A package for analysation of open beam files (OBP and OBF)."
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
dependencies = [
    "obplib",
    "numpy",
    "py3mf_slicer",
    "json",
    "csv",
    "matplotlib",
    "scipy",
    "pandas",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/wiberganton/OBAnalyser"
Repository = "https://github.com/wiberganton/OBAnalyser"